render Graph component into detail screen
